refactor openssa.l2.util.lm submodule
