+++ openssa/l2/reasoning/abstract.py
@@ -6,7 +6,7 @@ from dataclasses import dataclass, field
 from typing import TypeVar
 
 from openssa.l2.task.abstract import ATask
-from openssa.l2.llms import AnLLM, OpenAILLM
+from openssa.l2.util.lm import AnLM, OpenAILM
 
 
 @dataclass
@@ -14,7 +14,7 @@ class AbstractReasoner(ABC):
     """Abstract Reasoner."""
 
     # language model for reasoning
-    lm: AnLLM = field(default_factory=OpenAILLM.from_defaults)
+    lm: AnLM = field(default_factory=OpenAILM.from_defaults)
 
     @abstractmethod
     def reason(self, task: ATask, n_words: int = 1000) -> str:
